Flytter esbuild-scriptet til egen js-fil. Fikser problem der next.js server og custom server fikk separate cache-instanser.
