+++ server/package.json
@@ -3,7 +3,7 @@
     "version": "0.0.0",
     "private": true,
     "scripts": {
-        "build": "esbuild src/server.ts src/custom-cache-handler.ts --bundle --platform=node --packages=external --outdir=.dist --out-extension:.js=.cjs",
+        "build": "node build.mjs",
         "start": "node -r dotenv/config .dist/server.cjs dotenv_config_path=../.env",
         "nodemon-start": "nodemon -r dotenv/config -w .dist .dist/server.cjs dotenv_config_path=../.env",
         "dev": "npm run build && concurrently \"npm run build -- --watch\" \"npm run nodemon-start\"",
@@ -25,6 +25,7 @@
         "@types/on-headers": "1.0.3",
         "concurrently": "8.2.2",
         "esbuild": "0.20.0",
+        "esbuild-plugin-replace": "1.4.0",
         "express": "4.18.2",
         "nodemon": "3.0.2"
     }
